requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types @ file:///C:/b/abs_0dmaoyhhj3/croot/annotated-types_1709542968311/work
anthropic==0.49.0
anyio==4.8.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
async-timeout==4.0.3
attrs==25.1.0
backports.tarfile==1.2.0
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.1
certifi==2024.12.14
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
cloudpickle==3.1.1
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
contourpy==1.3.1
crashtest==0.4.1
cryptography @ file:///C:/b/abs_e2lzchf4i6/croot/cryptography_1732130411942/work
cycler==0.12.1
dashscope==1.22.1
dataclasses-json==0.6.7
debugpy @ file:///D:/bld/debugpy_1741148401445/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
distlib==0.3.9
distro @ file:///C:/b/abs_71xr36ua5r/croot/distro_1714488282676/work
dulwich==0.22.7
eclipse-sumo==1.21.0
exceptiongroup @ file:///C:/b/abs_c5h1o1_b5b/croot/exceptiongroup_1706031441653/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1733569351617/work
Farama-Notifications==0.0.4
fastapi==0.115.8
fastjsonschema==2.21.1
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
grpcio==1.70.0
gymnasium==1.0.0
h11 @ file:///C:/b/abs_1czwoyexjf/croot/h11_1706652332846/work
httpcore @ file:///C:/b/abs_55n7g233bw/croot/httpcore_1706728507241/work
httpx @ file:///C:/b/abs_43e135shby/croot/httpx_1723474830126/work
httpx-sse==0.4.0
idna==3.10
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
installer==0.7.0
ipykernel @ file:///D:/bld/ipykernel_1719845595208/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1741457807/work
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.5
jiter @ file:///C:/b/abs_acsxrw36a6/croot/jiter_1729808947806/work
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///D:/bld/jupyter_core_1727163532151/work
keyring==25.6.0
kiwisolver==1.4.8
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.43
langchain-text-splitters==0.3.6
langsmith==0.3.13
lap==0.5.12
libsumo==1.21.0
loguru==0.7.3
lxml==5.4.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mcp==1.6.0
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.25.1
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.4.2
numpy==2.2.1
openai @ file:///C:/b/abs_fbl6noof_d/croot/openai_1738233380947/work
opencv-python==4.10.0.84
orjson==3.10.15
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas==2.2.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
patsy==1.0.1
pettingzoo==1.24.3
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.0.0
pkginfo==1.12.0
platformdirs==4.3.6
plotly==6.0.0
plotly-express==0.4.1
poetry==2.0.1
poetry-core==2.0.1
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1737453357274/work
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
py-cpuinfo==9.0.0
pyarrow==19.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic @ file:///C:/b/abs_27dx58x550/croot/pydantic_1734736090499/work
pydantic-settings==2.8.1
pydantic_core @ file:///C:/b/abs_bdosz7qwys/croot/pydantic-core_1734726071532/work
pydeck==0.9.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
PyMySQL @ file:///C:/ci_310/pymysql_1643295949948/work
pyparsing==3.2.0
pyproject_hooks==1.2.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-dotenv==1.0.1
pytz==2024.2
pywin32==307
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq @ file:///D:/bld/pyzmq_1741805158588/work
RapidFuzz==3.12.1
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
SciencePlots==2.1.1
scipy==1.14.1
seaborn==0.13.2
shapely==2.0.6
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smmap==5.0.2
sniffio @ file:///C:/b/abs_3akdewudo_/croot/sniffio_1705431337396/work
SQLAlchemy==2.0.38
sse-starlette==2.2.1
stable_baselines3==2.5.0
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.45.3
statsmodels==0.14.4
streamlit==1.42.0
sumolib==1.22.0
SuperSuit==3.9.3
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensordict==0.7.2
tinyscaler==1.2.8
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.4.1
torch-directml==0.2.5.dev240914
torchrl==0.7.1
torchvision==0.19.1
tornado @ file:///D:/bld/tornado_1732615925919/work
tqdm @ file:///C:/b/abs_0eh9b6xugj/croot/tqdm_1738945553987/work
traci==1.22.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
trove-classifiers==2025.1.15.22
-e git+https://github.com/Traffic-Alpha/TransSimHub.git@00f1b84e2f16ff27d07d1d9e3a8f7d80a47ecfc5#egg=tshub
typing-inspect==0.9.0
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata==2024.2
ultralytics==8.3.55
ultralytics-thop==2.0.13
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.2
watchdog==6.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
websocket-client==1.8.0
Werkzeug==3.1.3
win32_setctime==1.2.0
yarl==1.18.3
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
zstandard==0.23.0
